requirements.txt
# --- Web framework ---
fastapi==0.116.1
starlette==0.47.2
uvicorn==0.35.0
httptools==0.6.4
uvloop==0.21.0
websockets==15.0.1
anyio==4.9.0
sniffio==1.3.1

# --- Seguridad ---
bcrypt==4.3.0
cryptography==45.0.5
rsa==4.9.1
ecdsa==0.19.1
cffi==1.17.1
pycparser==2.22
email_validator==2.2.0

# --- Base de datos ---
SQLAlchemy==2.0.41
sqlmodel==0.0.24
greenlet==3.2.3

# --- Desarrollo ---
black==25.1.0
mypy_extensions==1.1.0
pathspec==0.12.1
platformdirs==4.3.8
click==8.2.1
shellingham==1.5.4
typer==0.16.0
watchfiles==1.1.0

# --- CLI tools ---
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4

# --- HTTP cliente ---
httpx==0.28.1
httpcore==1.0.9
idna==3.10
urllib3==2.5.0
dnspython==2.7.0

# --- Validación ---
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
typing_extensions==4.14.1
typing-inspection==0.4.1

# --- Datos falsos ---
Faker==37.4.2

# --- otros ---
python-dotenv==1.1.1
PyYAML==6.0.2
six==1.17.0
certifi==2025.7.14
rignore==0.6.4
tzdata==2025.2
sentry-sdk==2.33.2
Jinja2==3.1.6
MarkupSafe==3.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.2
rich==14.0.0
rich-toolkit==0.14.8


